moved task.js to dummyData folder
